requirements.txt
flask
faiss-cpu
sentence_transformers
numpy